using wall bounding for isolated air particles in two pahse flow
