+++ src/shared/particle_dynamics/fluid_dynamics/distance_from_wall.cpp
@@ -1,0 +1,60 @@
+#include "distance_from_wall.h"
+
+namespace SPH
+{
+namespace fluid_dynamics
+{
+//=================================================================================================//
+DistanceFromWall::DistanceFromWall(BaseContactRelation &wall_contact_relation)
+    : LocalDynamics(wall_contact_relation.getSPHBody()), FSIContactData(wall_contact_relation),
+      spacing_ref_(sph_body_.sph_adaptation_->ReferenceSpacing()),
+      distance_default_(100.0 * spacing_ref_), distance_min_(0.25 * spacing_ref_),
+      pos_(particles_->pos_),
+      distance_from_wall_(*particles_->registerSharedVariable<Vecd>("DistanceFromWall"))
+{
+    for (size_t k = 0; k != contact_particles_.size(); ++k)
+    {
+        wall_pos_.push_back(&(contact_particles_[k]->pos_));
+        wall_n_.push_back(&(contact_particles_[k]->n_));
+        wall_phi_.push_back(contact_particles_[k]->getVariableByName<Real>("SignedDistance"));
+    }
+}
+//=================================================================================================//
+void DistanceFromWall::interaction(size_t index_i, Real dt)
+{
+    Vecd distance = distance_default_ * Vecd::Ones();
+    Vecd normal = Vecd::Ones();
+    for (size_t k = 0; k < contact_configuration_.size(); ++k)
+    {
+        StdLargeVec<Vecd> &pos_k = *(wall_pos_[k]);
+        StdLargeVec<Vecd> &n_k = *(wall_n_[k]);
+        StdLargeVec<Real> &phi_k = *(wall_phi_[k]);
+        Neighborhood &contact_neighborhood = (*contact_configuration_[k])[index_i];
+        for (size_t n = 0; n != contact_neighborhood.current_size_; ++n)
+        {
+            size_t index_j = contact_neighborhood.j_[n];
+            Vecd temp = (pos_[index_i] - pos_k[index_j]) + phi_k[index_j] * n_k[index_j];
+            if (temp.squaredNorm() < distance.squaredNorm())
+            {
+                distance = temp;       // more reliable distance
+                normal = n_k[index_j]; // more reliable normal
+            }
+        }
+    }
+
+    //prediction with regularization
+    Vecd normal_distance = distance.dot(normal) * normal;
+    Real limiter = SMIN(3.0 * (distance - normal_distance).norm() / spacing_ref_, 1.0);
+    distance_from_wall_[index_i] = (1.0 - limiter) * normal_distance + limiter * distance;
+
+    //bounding if the particle is too close to the wall
+    if (distance_from_wall_[index_i].dot(normal) < distance_min_)
+    {
+        pos_[index_i] -= distance_from_wall_[index_i]; // first moved to surface
+        distance_from_wall_[index_i] = 0.5 * spacing_ref_ * normal;
+        pos_[index_i] += distance_from_wall_[index_i]; // final bounded position
+    }
+}
+//=================================================================================================//
+} // namespace fluid_dynamics
+} // namespace SPH
