+++ src/shared/particle_dynamics/fluid_dynamics/velocity_gradient.cpp
@@ -32,45 +32,5 @@ void VelocityGradient<Contact<Wall>>::interaction(size_t index_i, Real dt)
     vel_grad_[index_i] += vel_grad;
 }
 //=================================================================================================//
-DistanceFromWall::DistanceFromWall(BaseContactRelation &wall_contact_relation)
-    : LocalDynamics(wall_contact_relation.getSPHBody()), FSIContactData(wall_contact_relation),
-      spacing_ref_(sph_body_.sph_adaptation_->ReferenceSpacing()),
-      distance_from_wall_(*particles_->registerSharedVariable<Vecd>("DistanceFromWall"))
-{
-    for (size_t k = 0; k != contact_particles_.size(); ++k)
-    {
-        wall_n_.push_back(&(contact_particles_[k]->n_));
-        wall_phi_.push_back(contact_particles_[k]->getVariableByName<Real>("SignedDistance"));
-    }
-}
-//=================================================================================================//
-void DistanceFromWall::interaction(size_t index_i, Real dt)
-{
-    Vecd distance = 100.0 * spacing_ref_ * Vecd::Ones();
-    Vecd normal = Vecd::Ones();
-    for (size_t k = 0; k < contact_configuration_.size(); ++k)
-    {
-        StdLargeVec<Vecd> &n_k = *(wall_n_[k]);
-        StdLargeVec<Real> &phi_k = *(wall_phi_[k]);
-        Neighborhood &contact_neighborhood = (*contact_configuration_[k])[index_i];
-        for (size_t n = 0; n != contact_neighborhood.current_size_; ++n)
-        {
-            size_t index_j = contact_neighborhood.j_[n];
-            const Vecd &e_ij = contact_neighborhood.e_ij_[n];
-
-            Vecd temp = contact_neighborhood.r_ij_[n] * e_ij + phi_k[index_j] * n_k[index_j];
-            if (temp.squaredNorm() < distance.squaredNorm())
-            {
-                distance = temp;       // more reliable distance
-                normal = n_k[index_j]; // more reliable normal
-            }
-        }
-    }
-
-    Vecd normal_distance = distance.dot(normal) * normal;
-    Real limiter = SMIN(3.0 * (distance - normal_distance).norm() / spacing_ref_, 1.0);
-    distance_from_wall_[index_i] = (1.0 - limiter) * normal_distance + limiter * distance;
-}
-//=================================================================================================//
 } // namespace fluid_dynamics
 } // namespace SPH
