+++ src/shared/particle_dynamics/fluid_dynamics/velocity_gradient.h
@@ -83,20 +83,6 @@ class VelocityGradient<Contact<Wall>> : public InteractionWithWall<VelocityGradi
 template <class KernelCorrectionType>
 using VelocityGradientWithWall = ComplexInteraction<VelocityGradient<Inner<KernelCorrectionType>, Contact<Wall>>>;
 
-class DistanceFromWall : public LocalDynamics, public FSIContactData
-{
-  public:
-    explicit DistanceFromWall(BaseContactRelation &wall_contact_relation);
-    virtual ~DistanceFromWall(){};
-    void interaction(size_t index_i, Real dt = 0.0);
-
-  protected:
-    Real spacing_ref_;
-    StdLargeVec<Vecd> &distance_from_wall_;
-    StdVec<StdLargeVec<Vecd> *> wall_n_;
-    StdVec<StdLargeVec<Real> *> wall_phi_;
-};
-
 } // namespace fluid_dynamics
 } // namespace SPH
 #endif // VELOCITY_GRADIENT_H
